pkg/syslinux: rip syslinux parser out of PXE
