+++ pkg/ipxe/ipxe.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// Package ipxe aims to implement a trivial IPXE configuration handler.
+// Package ipxe implements a trivial IPXE config file parser.
 package ipxe
 
 import (
@@ -14,7 +14,7 @@ import (
 	"strings"
 
 	"github.com/u-root/u-root/pkg/boot"
-	"github.com/u-root/u-root/pkg/pxe"
+	"github.com/u-root/u-root/pkg/syslinux"
 	"github.com/u-root/u-root/pkg/uio"
 )
 
@@ -24,39 +24,39 @@ var (
 	ErrNotIpxeScript = errors.New("config file is not ipxe as it does not start with #!ipxe")
 )
 
-// Config encapsulates a parsed ipxe configuration file.
+// parser encapsulates a parsed ipxe configuration file.
 //
 // We currently only support kernel and initrd commands.
-type Config struct {
-	BootImage *boot.LinuxImage
+type parser struct {
+	bootImage *boot.LinuxImage
 
-	schemes pxe.Schemes
+	schemes syslinux.Schemes
 }
 
-// NewConfig returns a new IPXE configuration with the file at URL and default
+// ParseConfig returns a new  configuration with the file at URL and default
 // schemes.
 //
-// See NewConfigWithSchemes for more details.
-func NewConfig(configURL *url.URL) (*Config, error) {
-	return NewConfigWithSchemes(configURL, pxe.DefaultSchemes)
+// See ParseConfigWithSchemes for more details.
+func ParseConfig(configURL *url.URL) (*boot.LinuxImage, error) {
+	return ParseConfigWithSchemes(configURL, syslinux.DefaultSchemes)
 }
 
-// NewConfigWithSchemes returns a new IPXE configuration with the file at URL
+// ParseConfigWithSchemes returns a new  configuration with the file at URL
 // and schemes `s`.
 //
 // `s` is used to get files referred to by URLs in the configuration.
-func NewConfigWithSchemes(configURL *url.URL, s pxe.Schemes) (*Config, error) {
-	c := &Config{
+func ParseConfigWithSchemes(configURL *url.URL, s syslinux.Schemes) (*boot.LinuxImage, error) {
+	c := &parser{
 		schemes: s,
 	}
 	if err := c.getAndParseFile(configURL); err != nil {
 		return nil, err
 	}
-	return c, nil
+	return c.bootImage, nil
 }
 
 // getAndParse parses the config file downloaded from `url` and fills in `c`.
-func (c *Config) getAndParseFile(u *url.URL) error {
+func (c *parser) getAndParseFile(u *url.URL) error {
 	r, err := c.schemes.LazyGetFile(u)
 	if err != nil {
 		return err
@@ -74,20 +74,19 @@ func (c *Config) getAndParseFile(u *url.URL) error {
 }
 
 // getFile parses `surl` and returns an io.Reader for the requested url.
-func (c *Config) getFile(surl string) (io.ReaderAt, error) {
+func (c *parser) getFile(surl string) (io.ReaderAt, error) {
 	u, err := url.Parse(surl)
 	if err != nil {
 		return nil, fmt.Errorf("could not parse URL %q: %v", surl, err)
 	}
-
 	return c.schemes.LazyGetFile(u)
 }
 
 // parseIpxe parses `config` and constructs a BootImage for `c`.
-func (c *Config) parseIpxe(config string) error {
+func (c *parser) parseIpxe(config string) error {
 	// A trivial ipxe script parser.
 	// Currently only supports kernel and initrd commands.
-	c.BootImage = &boot.LinuxImage{}
+	c.bootImage = &boot.LinuxImage{}
 
 	for _, line := range strings.Split(config, "\n") {
 		// Skip blank lines and comment lines.
@@ -98,7 +97,7 @@ func (c *Config) parseIpxe(config string) error {
 
 		args := strings.Fields(line)
 		if len(args) <= 1 {
-			log.Printf("Ignoring unsupported ipxe cmd: %s\n", line)
+			log.Printf("Ignoring unsupported ipxe cmd: %s", line)
 			continue
 		}
 		cmd := strings.ToLower(args[0])
@@ -109,11 +108,11 @@ func (c *Config) parseIpxe(config string) error {
 			if err != nil {
 				return err
 			}
-			c.BootImage.Kernel = k
+			c.bootImage.Kernel = k
 
 			// Add cmdline if there are any.
 			if len(args) > 2 {
-				c.BootImage.Cmdline = strings.Join(args[2:], " ")
+				c.bootImage.Cmdline = strings.Join(args[2:], " ")
 			}
 
 		case "initrd":
@@ -121,10 +120,10 @@ func (c *Config) parseIpxe(config string) error {
 			if err != nil {
 				return err
 			}
-			c.BootImage.Initrd = i
+			c.bootImage.Initrd = i
 
 		default:
-			log.Printf("Ignoring unsupported ipxe cmd: %s\n", line)
+			log.Printf("Ignoring unsupported ipxe cmd: %s", line)
 		}
 	}
