+++ pkg/ipxe/ipxe_test.go
@@ -6,39 +6,52 @@ package ipxe
 
 import (
 	"fmt"
+	"io"
 	"net/url"
 	"reflect"
+	"strings"
 	"testing"
 
-	"github.com/u-root/u-root/pkg/pxe"
+	"github.com/u-root/u-root/pkg/boot"
+	"github.com/u-root/u-root/pkg/syslinux"
 	"github.com/u-root/u-root/pkg/uio"
 )
+
+func mustReadAll(r io.ReaderAt) string {
+	if r == nil {
+		return ""
+	}
+	b, err := uio.ReadAll(r)
+	if err != nil {
+		return fmt.Sprintf("read error: %s", err)
+	}
+	return string(b)
+}
+
+type errorReader struct {
+	err error
+}
+
+func (e errorReader) ReadAt(p []byte, n int64) (int, error) {
+	return 0, e.err
+}
 
 func TestIpxeConfig(t *testing.T) {
 	content1 := "1111"
 	content2 := "2222"
 
-	type config struct {
-		kernel    string
-		kernelErr error
-		initrd    string
-		initrdErr error
-		cmdline   string
-	}
-
 	for i, tt := range []struct {
 		desc       string
-		schemeFunc func() pxe.Schemes
+		schemeFunc func() syslinux.Schemes
 		curl       *url.URL
-		config     *Config
-		want       config
+		want       *boot.LinuxImage
 		err        error
 	}{
 		{
 			desc: "all files exist, simple config with no cmdline",
-			schemeFunc: func() pxe.Schemes {
-				s := make(pxe.Schemes)
-				fs := pxe.NewMockScheme("http")
+			schemeFunc: func() syslinux.Schemes {
+				s := make(syslinux.Schemes)
+				fs := syslinux.NewMockScheme("http")
 				conf := `#!ipxe
 				kernel http://someplace.com/foobar/pxefiles/kernel
 				initrd http://someplace.com/foobar/pxefiles/initrd
@@ -54,16 +67,16 @@ func TestIpxeConfig(t *testing.T) {
 				Host:   "someplace.com",
 				Path:   "/foobar/pxefiles/ipxeconfig",
 			},
-			want: config{
-				kernel: content1,
-				initrd: content2,
+			want: &boot.LinuxImage{
+				Kernel: strings.NewReader(content1),
+				Initrd: strings.NewReader(content2),
 			},
 		},
 		{
 			desc: "all files exist, simple config, no initrd",
-			schemeFunc: func() pxe.Schemes {
-				s := make(pxe.Schemes)
-				fs := pxe.NewMockScheme("http")
+			schemeFunc: func() syslinux.Schemes {
+				s := make(syslinux.Schemes)
+				fs := syslinux.NewMockScheme("http")
 				conf := `#!ipxe
 				kernel http://someplace.com/foobar/pxefiles/kernel
 				boot`
@@ -77,15 +90,15 @@ func TestIpxeConfig(t *testing.T) {
 				Host:   "someplace.com",
 				Path:   "/foobar/pxefiles/ipxeconfig",
 			},
-			want: config{
-				kernel: content1,
+			want: &boot.LinuxImage{
+				Kernel: strings.NewReader(content1),
 			},
 		},
 		{
 			desc: "comments and blank lines",
-			schemeFunc: func() pxe.Schemes {
-				s := make(pxe.Schemes)
-				fs := pxe.NewMockScheme("http")
+			schemeFunc: func() syslinux.Schemes {
+				s := make(syslinux.Schemes)
+				fs := syslinux.NewMockScheme("http")
 				conf := `#!ipxe
 				# the next line is blank
 
@@ -101,15 +114,15 @@ func TestIpxeConfig(t *testing.T) {
 				Host:   "someplace.com",
 				Path:   "/foobar/pxefiles/ipxeconfig",
 			},
-			want: config{
-				kernel: content1,
+			want: &boot.LinuxImage{
+				Kernel: strings.NewReader(content1),
 			},
 		},
 		{
 			desc: "kernel does not exist, simple config",
-			schemeFunc: func() pxe.Schemes {
-				s := make(pxe.Schemes)
-				fs := pxe.NewMockScheme("http")
+			schemeFunc: func() syslinux.Schemes {
+				s := make(syslinux.Schemes)
+				fs := syslinux.NewMockScheme("http")
 				conf := `#!ipxe
 				kernel http://someplace.com/foobar/pxefiles/kernel
 				boot`
@@ -122,24 +135,24 @@ func TestIpxeConfig(t *testing.T) {
 				Host:   "someplace.com",
 				Path:   "/foobar/pxefiles/ipxeconfig",
 			},
-			want: config{
-				kernelErr: &pxe.URLError{
+			want: &boot.LinuxImage{
+				Kernel: errorReader{&syslinux.URLError{
 					URL: &url.URL{
 						Scheme: "http",
 						Host:   "someplace.com",
 						Path:   "/foobar/pxefiles/kernel",
 					},
-					Err: pxe.ErrNoSuchFile,
-				},
-				initrd:  "",
-				cmdline: "",
+					Err: syslinux.ErrNoSuchFile,
+				}},
+				Initrd:  nil,
+				Cmdline: "",
 			},
 		},
 		{
 			desc: "config file does not exist",
-			schemeFunc: func() pxe.Schemes {
-				s := make(pxe.Schemes)
-				fs := pxe.NewMockScheme("http")
+			schemeFunc: func() syslinux.Schemes {
+				s := make(syslinux.Schemes)
+				fs := syslinux.NewMockScheme("http")
 				s.Register(fs.Scheme, fs)
 				return s
 			},
@@ -148,20 +161,20 @@ func TestIpxeConfig(t *testing.T) {
 				Host:   "someplace.com",
 				Path:   "/foobar/pxefiles/ipxeconfig",
 			},
-			err: &pxe.URLError{
+			err: &syslinux.URLError{
 				URL: &url.URL{
 					Scheme: "http",
 					Host:   "someplace.com",
 					Path:   "/foobar/pxefiles/ipxeconfig",
 				},
-				Err: pxe.ErrNoSuchHost,
+				Err: syslinux.ErrNoSuchHost,
 			},
 		},
 		{
 			desc: "invalid config",
-			schemeFunc: func() pxe.Schemes {
-				s := make(pxe.Schemes)
-				fs := pxe.NewMockScheme("http")
+			schemeFunc: func() syslinux.Schemes {
+				s := make(syslinux.Schemes)
+				fs := syslinux.NewMockScheme("http")
 				fs.Add("someplace.com", "/foobar/pxefiles/ipxeconfig", "")
 				s.Register(fs.Scheme, fs)
 				return s
@@ -175,9 +188,9 @@ func TestIpxeConfig(t *testing.T) {
 		},
 		{
 			desc: "empty config",
-			schemeFunc: func() pxe.Schemes {
-				s := make(pxe.Schemes)
-				fs := pxe.NewMockScheme("http")
+			schemeFunc: func() syslinux.Schemes {
+				s := make(syslinux.Schemes)
+				fs := syslinux.NewMockScheme("http")
 				conf := `#!ipxe`
 				fs.Add("someplace.com", "/foobar/pxefiles/ipxeconfig", conf)
 				s.Register(fs.Scheme, fs)
@@ -188,13 +201,13 @@ func TestIpxeConfig(t *testing.T) {
 				Host:   "someplace.com",
 				Path:   "/foobar/pxefiles/ipxeconfig",
 			},
-			want: config{},
+			want: &boot.LinuxImage{},
 		},
 		{
 			desc: "valid config with kernel cmdline args",
-			schemeFunc: func() pxe.Schemes {
-				s := make(pxe.Schemes)
-				fs := pxe.NewMockScheme("http")
+			schemeFunc: func() syslinux.Schemes {
+				s := make(syslinux.Schemes)
+				fs := syslinux.NewMockScheme("http")
 				conf := `#!ipxe
 				kernel http://someplace.com/foobar/pxefiles/kernel earlyprintk=ttyS0 printk=ttyS0
 				boot`
@@ -208,27 +221,27 @@ func TestIpxeConfig(t *testing.T) {
 				Host:   "someplace.com",
 				Path:   "/foobar/pxefiles/ipxeconfig",
 			},
-			want: config{
-				kernel:  content1,
-				cmdline: "earlyprintk=ttyS0 printk=ttyS0",
+			want: &boot.LinuxImage{
+				Kernel:  strings.NewReader(content1),
+				Cmdline: "earlyprintk=ttyS0 printk=ttyS0",
 			},
 		},
 		{
 			desc: "multi-scheme valid config",
-			schemeFunc: func() pxe.Schemes {
+			schemeFunc: func() syslinux.Schemes {
 				conf := `#!ipxe
 				kernel tftp://1.2.3.4/foobar/pxefiles/kernel
                                 initrd http://someplace.com/someinitrd.gz
 				boot`
 
-				tftp := pxe.NewMockScheme("tftp")
+				tftp := syslinux.NewMockScheme("tftp")
 				tftp.Add("1.2.3.4", "/foobar/pxefiles/kernel", content1)
 
-				http := pxe.NewMockScheme("http")
+				http := syslinux.NewMockScheme("http")
 				http.Add("someplace.com", "/foobar/pxefiles/ipxeconfig", conf)
 				http.Add("someplace.com", "/someinitrd.gz", content2)
 
-				s := make(pxe.Schemes)
+				s := make(syslinux.Schemes)
 				s.Register(tftp.Scheme, tftp)
 				s.Register(http.Scheme, http)
 				return s
@@ -238,16 +251,16 @@ func TestIpxeConfig(t *testing.T) {
 				Host:   "someplace.com",
 				Path:   "/foobar/pxefiles/ipxeconfig",
 			},
-			want: config{
-				kernel: content1,
-				initrd: content2,
+			want: &boot.LinuxImage{
+				Kernel: strings.NewReader(content1),
+				Initrd: strings.NewReader(content2),
 			},
 		},
 		{
 			desc: "valid config with unsupported cmds",
-			schemeFunc: func() pxe.Schemes {
-				s := make(pxe.Schemes)
-				fs := pxe.NewMockScheme("http")
+			schemeFunc: func() syslinux.Schemes {
+				s := make(syslinux.Schemes)
+				fs := syslinux.NewMockScheme("http")
 				conf := `#!ipxe
 				kernel http://someplace.com/foobar/pxefiles/kernel
                                 initrd http://someplace.com/someinitrd.gz
@@ -264,55 +277,33 @@ func TestIpxeConfig(t *testing.T) {
 				Host:   "someplace.com",
 				Path:   "/foobar/pxefiles/ipxeconfig",
 			},
-			want: config{
-				kernel: content1,
-				initrd: content2,
+			want: &boot.LinuxImage{
+				Kernel: strings.NewReader(content1),
+				Initrd: strings.NewReader(content2),
 			},
 		},
 	} {
 		t.Run(fmt.Sprintf("Test [%02d] %s", i, tt.desc), func(t *testing.T) {
-			c, err := NewConfigWithSchemes(tt.curl, tt.schemeFunc())
+			got, err := ParseConfigWithSchemes(tt.curl, tt.schemeFunc())
 			if !reflect.DeepEqual(err, tt.err) {
 				t.Errorf("NewConfigWithSchemes() got %v, want %v", err, tt.err)
 				return
 			} else if err != nil {
 				return
 			}
-
-			got := c.BootImage
 			want := tt.want
 
 			// Same kernel?
-			if got.Kernel == nil && (len(want.kernel) > 0 || want.kernelErr != nil) {
-				t.Errorf("want kernel, got none")
+			if !uio.ReaderAtEqual(got.Kernel, want.Kernel) {
+				t.Errorf("got kernel %s, want %s", mustReadAll(got.Kernel), mustReadAll(want.Kernel))
 			}
-			if got.Kernel != nil {
-				k, err := uio.ReadAll(got.Kernel)
-				if !reflect.DeepEqual(err, want.kernelErr) {
-					t.Errorf("could not read kernel. got: %v, want %v", err, want.kernelErr)
-				}
-				if got, want := string(k), want.kernel; got != want {
-					t.Errorf("got kernel %s, want %s", got, want)
-				}
-			}
-
 			// Same initrd?
-			if got.Initrd == nil && (len(want.initrd) > 0 || want.initrdErr != nil) {
-				t.Errorf("want initrd, got none")
+			if !uio.ReaderAtEqual(got.Initrd, want.Initrd) {
+				t.Errorf("got initrd %s, want %s", mustReadAll(got.Initrd), mustReadAll(want.Initrd))
 			}
-			if got.Initrd != nil {
-				i, err := uio.ReadAll(got.Initrd)
-				if err != want.initrdErr {
-					t.Errorf("could not read initrd. got: %v, want %v", err, want.initrdErr)
-				}
-				if got, want := string(i), want.initrd; got != want {
-					t.Errorf("got initrd %s, want %s", got, want)
-				}
-			}
-
 			// Same cmdline?
-			if got, want := got.Cmdline, want.cmdline; got != want {
-				t.Errorf("got cmdline %s, want %s", got, want)
+			if got.Cmdline != want.Cmdline {
+				t.Errorf("got cmdline %s, want %s", got.Cmdline, want.Cmdline)
 			}
 		})
 	}
